refactor(visualization): extract file loaders' shared helpers/constants to low layers
Structural prep (Tidy First) before turning the loaders into a feature:

- Extract CC_FILE_EXTENSION (".cc.json") from a static on LoadFileService into
  model/files/files.ts, so util/fileNameHelper and util/fileDownloader stop
  importing the service for a constant (removes a util->service inversion).
- Extract the MetricQueryParemter enum out of updateQueryParameters.effect into
  a sibling metricQueryParameter.ts; loadInitialFile.service imported the enum
  from the effect while the effect injects the service — extracting it breaks
  that cycle.
- Move the pure, ngrx-free helpers fileParser, fileRoot, and urlExtractor (+spec)
  from services/ into util/, where their util/state/feature consumers can use
  them without a layering inversion.

No behavior change. tsc, full Jest suite (378 suites), and dep-cruiser
(0 errors) pass.
