+++ visualization/app/codeCharta/features/globalSettings/components/globalConfigurationDialog/confirmResetMapDialog/confirmResetMapDialog.spec.ts
@@ -5,10 +5,10 @@ import { MockStore, provideMockStore } from "@ngrx/store/testing"
 import { render, screen } from "@testing-library/angular"
 import userEvent from "@testing-library/user-event"
 import "fake-indexeddb/auto"
-import { getNameDataPair } from "../../../../../services/loadFile/fileParser"
+import { getNameDataPair } from "../../../../../util/fileParser"
 import { LoadFileService } from "../../../../../services/loadFile/loadFile.service"
 import { LoadInitialFileService, sampleFile1, sampleFile2 } from "../../../../../services/loadInitialFile/loadInitialFile.service"
-import { UrlExtractor } from "../../../../../services/loadInitialFile/urlExtractor"
+import { UrlExtractor } from "../../../../../util/urlExtractor"
 import * as resetChosenMetricsEffect from "../../../../../state/effects/resetChosenMetrics/resetChosenMetrics.effect"
 import { metricDataSelector } from "../../../../../state/selectors/accumulatedData/metricData/metricData.selector"
 import { setState } from "../../../../../state/store/state.actions"
@@ -18,7 +18,7 @@ import * as indexedDBWriter from "../../../../../util/indexedDB/indexedDBWriter"
 import { ConfirmResetMapDialogComponent } from "./confirmResetMapDialog.component"
 
 jest.mock("../../../../../util/indexedDB/indexedDBWriter")
-jest.mock("../../../../../services/loadInitialFile/urlExtractor")
+jest.mock("../../../../../util/urlExtractor")
 jest.mock("../../../../../state/effects/resetChosenMetrics/resetChosenMetrics.effect")
 
 describe("ConfirmResetMapDialogComponent", () => {
