+++ visualization/app/codeCharta/model/files/files.ts
@@ -1,5 +1,7 @@
 import { CCFile } from "../../codeCharta.model"
 
+export const CC_FILE_EXTENSION = ".cc.json"
+
 export interface FileState {
     file: CCFile
     selectedAs: FileSelectionState
