+++ visualization/app/codeCharta/services/loadInitialFile/loadInitialFile.service.ts
@@ -9,14 +9,14 @@ import { ExportCCFile } from "../../codeCharta.api.model"
 import { AppSettings, CcState, DynamicSettings, FileSettings, NameDataPair } from "../../codeCharta.model"
 import { FileState } from "../../model/files/files"
 import { getCCFiles } from "../../model/files/files.helper"
-import { MetricQueryParemter } from "../../state/effects/updateQueryParameters/updateQueryParameters.effect"
+import { MetricQueryParemter } from "../../state/effects/updateQueryParameters/metricQueryParameter"
 import { metricDataSelector } from "../../state/selectors/accumulatedData/metricData/metricData.selector"
 import { setDelta, setFiles } from "../../state/store/files/files.actions"
 import { ErrorDialogService } from "../../features/shared/components/errorDialog/errorDialog.service"
 import { buildHtmlMessage } from "../../util/loadFilesValidationToErrorDialog"
-import { getNameDataPair } from "../loadFile/fileParser"
+import { getNameDataPair } from "../../util/fileParser"
 import { LoadFileService, NO_FILES_LOADED_ERROR_MESSAGE } from "../loadFile/loadFile.service"
-import { UrlExtractor } from "./urlExtractor"
+import { UrlExtractor } from "../../util/urlExtractor"
 import { setCurrentFilesAreSampleFiles } from "../../state/store/appStatus/currentFilesAreSampleFiles/currentFilesAreSampleFiles.actions"
 import { setShowOutgoingEdges } from "../../state/store/appSettings/showEdges/outgoing/showOutgoingEdges.actions"
 import { setShowIncomingEdges } from "../../state/store/appSettings/showEdges/incoming/showIncomingEdges.actions"
