+++ visualization/app/codeCharta/state/effects/updateQueryParameters/metricQueryParameter.ts
@@ -1,0 +1,7 @@
+export enum MetricQueryParemter {
+    areaMetric = "area",
+    heightMetric = "height",
+    colorMetric = "color",
+    edgeMetric = "edge",
+    currentFilesAreSampleFiles = "currentFilesAreSampleFiles"
+}
